% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{all_mgi}
\alias{all_mgi}
\title{All MGI gene symbols}
\format{
An object of class \code{character} of length 48440.
}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
listMarts(host="www.ensembl.org")
mouse <- useMart(host="www.ensembl.org", "ENSEMBL_MART_ENSEMBL", 
                      dataset="mmusculus_gene_ensembl")
mgi_symbols = getBM(attributes=c("mgi_symbol","ensembl_gene_id"), 
                        mart=mouse)
all_mgi = unique(mgi_symbols$mgi_symbol)
save(all_mgi,file="all_mgi.Rda")
}
}
\usage{
all_mgi
}
\description{
A dataset containing all MGI symbols
}
\keyword{datasets}
