% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{all_mgi_wtEnsembl}
\alias{all_mgi_wtEnsembl}
\title{All MGI gene symbols with ENSEMBL gene IDs}
\format{
An object of class \code{data.frame} with 53952 rows and 2 columns.
}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
listMarts(host="www.ensembl.org")
mouse <- useMart(host="www.ensembl.org", "ENSEMBL_MART_ENSEMBL", 
                     dataset="mmusculus_gene_ensembl")
all_mgi_wtEnsembl = getBM(attributes=c("mgi_symbol","ensembl_gene_id"), 
                               mart=mouse)
save(all_mgi_wtEnsembl,file="all_mgi_wtEnsembl.Rda")
}
}
\usage{
all_mgi_wtEnsembl
}
\description{
A dataset containing all MGI symbols in first column, then ensembl_gene_id 
in second column
}
\keyword{datasets}
