% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{alzh_gwas_top100}
\alias{alzh_gwas_top100}
\title{Alzheimers disease top100 GWAS genes generated by MAGMA using the iGAP summary statistics}
\format{
An object of class \code{character} of length 100.
}
\source{
hgnc_symbols = getBM(attributes=c("hgnc_symbol","ensembl_gene_id","entrezgene"), mart=human)
magma2=merge(magma,hgnc_symbols,by="GENE")
write.csv(magma2[1:100,]$hgnc_symbol,file="Alzh_IGAP_top100magma.txt",quote=FALSE,row.names = FALSE)
}
\usage{
alzh_gwas_top100
}
\description{
Alzheimers disease top100 GWAS genes generated by MAGMA using the iGAP summary statistics
}
\keyword{datasets}
