% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cores.r
\name{assign_cores}
\alias{assign_cores}
\title{Assign cores}
\usage{
assign_cores(worker_cores = 0.9, verbose = TRUE)
}
\arguments{
\item{worker_cores}{Number (>1) or proportion (<1) of worker cores to use.}

\item{verbose}{Print messages.}
}
\value{
List of core allocations.
}
\description{
Assign cores automatically for parallel processing, while reserving some.
}
\keyword{internal}
