% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.specificity.into.quantiles.r
\name{bin.specificity.into.quantiles}
\alias{bin.specificity.into.quantiles}
\title{bin.specificity.into.quantiles}
\usage{
bin.specificity.into.quantiles(ctdIN, numberOfBins)
}
\arguments{
\item{ctdIN}{A single annotLevel of a ctd, i.e. ctd[[1]] (the function is 
intended to be used via apply)}

\item{numberOfBins}{Number of quantile 'bins' to use (40 is recommended)}
}
\value{
A ctd with $specificity_quantiles
}
\description{
\code{bin.specificity.into.quantiles} is an internal function used to convert
add '$specificity_quantiles' to a ctd
}
\examples{
data(ctd, package="ewceData")
ctd <- lapply(ctd, bin.specificity.into.quantiles, numberOfBins = 40)
print(ctd[[1]]$specificity_quantiles[1:3, ])
}
