% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.enrichment.test.r
\name{bootstrap.enrichment.test}
\alias{bootstrap.enrichment.test}
\title{Bootstrap celltype enrichment test}
\usage{
bootstrap.enrichment.test(
  sct_data = NA,
  hits = NA,
  bg = NA,
  genelistSpecies = "mouse",
  sctSpecies = "mouse",
  reps = 100,
  annotLevel = 1,
  geneSizeControl = FALSE,
  controlledCT = NULL
)
}
\arguments{
\item{sct_data}{List generated using \code{\link{generate.celltype.data}}}

\item{hits}{Array of MGI gene symbols containing the target gene list. 
Must be HGNC symbols if geneSizeControl=TRUE}

\item{bg}{Array of MGI gene symbols containing the background gene list. 
Must be HGNC symbols if geneSizeControl=TRUE}

\item{genelistSpecies}{Either 'mouse' or 'human' depending on whether MGI 
or HGNC symbols are used for gene lists}

\item{sctSpecies}{Either 'mouse' or 'human' depending on whether MGI or 
HGNC symbols are used for the single cell dataset}

\item{reps}{Number of random gene lists to generate (default=100 but should 
be over 10000 for publication quality results)}

\item{annotLevel}{an integer indicating which level of the annotation to 
analyse. Default = 1.}

\item{geneSizeControl}{a logical indicating whether you want to control for 
GC content and transcript length. Recommended if the gene list originates 
from genetic studies. Default is FALSE. If set to TRUE then human gene lists 
should be used rather than mouse.}

\item{controlledCT}{(optional) If not NULL, and instead is the name of a 
cell type, then the bootstrapping controls for expression within that 
cell type}
}
\value{
A list containing three data frames:
\itemize{
  \item \code{results}: dataframe in which each row gives the statistics 
  (p-value, fold change and number of standard deviations from the mean) 
  associated with the enrichment of the stated cell type in the gene list
  \item \code{hit.cells}: vector containing the summed proportion of 
  expression in each cell type for the target list
  \item \code{bootstrap_data}: matrix in which each row represents the 
  summed proportion of expression in each cell type for one of the 
  random lists
}
}
\description{
\code{bootstrap.enrichment.test} takes a genelist and a single cell type 
transcriptome dataset and determines the probability of enrichment and fold 
changes for each cell type.
}
\examples{
# Load the single cell data
data("ctd", package="ewceData")

# Set the parameters for the analysis
# Use 100 bootstrap lists for speed, for publishable analysis use >10000
reps <- 100 

# Load the gene list and get human orthologs
data("example_genelist", package="ewceData")
data("mouse_to_human_homologs", package="ewceData")
m2h <- unique(mouse_to_human_homologs[, c("HGNC.symbol", "MGI.symbol")])
mouse.hits <- 
    unique(m2h[m2h$HGNC.symbol \%in\% example_genelist, "MGI.symbol"])
human.hits <- 
    unique(m2h[m2h$HGNC.symbol \%in\% example_genelist, "HGNC.symbol"])
human.bg <- unique(m2h$HGNC.symbol)
mouse.bg <- unique(m2h$MGI.symbol)

# Bootstrap significance test, no control for transcript length or GC content
full_results <- bootstrap.enrichment.test(
    sct_data = ctd, hits = mouse.hits,
    bg = mouse.bg, reps = reps, annotLevel = 2, sctSpecies = "mouse", 
    genelistSpecies = "mouse"
)

# Bootstrap significance test control for transcript length and GC content
full_results <- bootstrap.enrichment.test(
    sct_data = ctd, hits = human.hits,
    bg = human.bg, reps = reps, annotLevel = 2, geneSizeControl = TRUE,
    sctSpecies = "mouse", genelistSpecies = "human"
)
}
