% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_plot.R
\name{bootstrap_plot}
\alias{bootstrap_plot}
\title{Bootstrap plot}
\usage{
bootstrap_plot(exp_mats, hit.exp, cc, savePath, listFileName, show_plot = TRUE)
}
\arguments{
\item{exp_mats}{exp_mats (#fix).}

\item{hit.exp}{hit.exp (#fix).}

\item{cc}{Cell-types to plot.}

\item{savePath}{Path to save plots to.}

\item{listFileName}{listFileName}

\item{show_plot}{Print the plot.}
}
\value{
Null output.
}
\description{
Plot bootstrap enrichment results.
}
