% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_quantiles.R
\name{calc_quantiles}
\alias{calc_quantiles}
\title{Calculate quantiles}
\usage{
calc_quantiles(vec, n_quantiles = 10, verbose = TRUE)
}
\arguments{
\item{vec}{Numeric vector.}

\item{n_quantiles}{Number of quantile bins to use.
Defaults to deciles (\code{n_quantiles=10}).}

\item{verbose}{Print messages.}
}
\value{
Quantiles.
}
\description{
Calculate quantiles
}
\keyword{internal}
