% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_celltype_specificity.r
\name{calculate_celltype_specificity}
\alias{calculate_celltype_specificity}
\title{Calculate celltype specificity}
\usage{
calculate_celltype_specificity(exp, annot, thresh = 0, trim = 0)
}
\arguments{
\item{exp}{A matrix containing the expression data, wherein rows are genes and columns are cells}

\item{annot}{A list of annotation vectors, where each vector has a name for each cell}

\item{thresh}{Expression threshold value below which a gene is dropped}

\item{trim}{Value determining how the trimmed mean is calculated (range: 0 to 0.5)}
}
\value{
A list array, in which each cell contains two data frames:
\itemize{
  \item \code{mean_exp}: stores the average value for each subcell type found across cells
  \item \code{specificity}: proportion of expression for each gene in each cell type
}
}
\description{
\code{calculate_celltype_specificity} takes expression & annotation data and determines
for each gene, the proportion of expression found in each celltype
}
\examples{
# Load the single cell data
data("cortex_mrna")
ctd = calculate_celltype_specificity(cortex_mrna$exp,cortex_mrna$annot$level2class)
}
