% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_meanexp_for_level.R
\name{calculate_meanexp_for_level}
\alias{calculate_meanexp_for_level}
\title{calculate_meanexp_for_level}
\usage{
calculate_meanexp_for_level(
  ctd_oneLevel,
  expMatrix,
  as_sparse = TRUE,
  verbose = TRUE
)
}
\value{
One level of a CellTypeDataset.
}
\description{
calculate_meanexp_for_level
}
\keyword{internal}
