% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_specificity_for_level.R
\name{calculate_specificity_for_level}
\alias{calculate_specificity_for_level}
\title{Calculate specificity for one CTD level}
\usage{
calculate_specificity_for_level(
  ctd_oneLevel,
  matrix_name = "mean_exp",
  as_sparse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{ctd_oneLevel}{One level from a CTD.}

\item{matrix_name}{Name of the matrix to extract.}

\item{as_sparse}{Whether to convert \code{exp} to sparse matrix}

\item{verbose}{Print messages.}
}
\value{
One CTD level.
}
\description{
Calculate specificity for one CellTypeDataset (CTD) level.
}
\keyword{internal}
