% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_list_dist.r
\name{cell_list_dist}
\alias{cell_list_dist}
\title{cell_list_dist}
\usage{
cell_list_dist(hitGenes, sct_data, annotLevel)
}
\arguments{
\item{hitGenes}{List of gene symbols containing the target gene list.}

\item{sct_data}{List generated using \link[EWCE]{generate_celltype_data}.}

\item{annotLevel}{An integer indicating which level of \code{sct_data} to
analyse (\emph{Default: 1}).}
}
\value{
Output (#fix).
}
\description{
specificity is generated in the main_CellTypeAnalysis_Preperation.r file
}
\keyword{internal}
