% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.ewce.genelist.input.r
\name{check.ewce.genelist.inputs}
\alias{check.ewce.genelist.inputs}
\title{check.ewce.genelist.inputs}
\usage{
check.ewce.genelist.inputs(
  sct_data,
  hits,
  bg,
  genelistSpecies,
  sctSpecies,
  geneSizeControl = FALSE
)
}
\arguments{
\item{sct_data}{List generated using \code{\link{generate.celltype.data}}}

\item{hits}{Array of MGI/HGNC gene symbols containing the target gene list.}

\item{bg}{Array of MGI/HGNC gene symbols containing the background gene list.}

\item{genelistSpecies}{Either 'mouse' or 'human' depending on whether MGI 
or HGNC symbols are used for gene lists}

\item{sctSpecies}{Either 'mouse' or 'human' depending on whether MGI 
or HGNC symbols are used for the single cell dataset}

\item{geneSizeControl}{Will genelists sampled control for GC content and 
transcript length? Boolean.}
}
\value{
A list containing
\itemize{
  \item \code{hits}: Array of MGI/HGNC gene symbols containing the target 
  gene list.
  \item \code{bg}: Array of MGI/HGNC gene symbols containing the background 
  gene list.
}
}
\description{
\code{check.ewce.genelist.inputs} Is used to check that hits and bg gene 
lists passed to EWCE are setup correctly. Checks they are the 
appropriate length.
Checks all hits genes are in bg. Checks the species match and if not 
reduces to 1:1 orthologs.
}
\examples{
library(ewceData)
# Called from "bootstrap.enrichment.test()" and "generate.bootstrap.plots()"
ctd <- ctd()
example_genelist <- example_genelist()
mouse_to_human_homologs <- mouse_to_human_homologs()
m2h <- unique(mouse_to_human_homologs[, c("HGNC.symbol", "MGI.symbol")])
mouse.hits <- 
     unique(m2h[m2h$HGNC.symbol \%in\% example_genelist, "MGI.symbol"])
mouse.bg <- unique(m2h$MGI.symbol)
checkedLists <- check.ewce.genelist.inputs(
    sct_data = ctd, hits = mouse.hits,
    bg = mouse.bg, genelistSpecies = "mouse", sctSpecies = "mouse"
)
}
