% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_annotLevels.R
\name{check_annotLevels}
\alias{check_annotLevels}
\title{check_annotLevels
 
First, check the number of annotations equals the number of columns
in the expression data.}
\usage{
check_annotLevels(annotLevels, exp)
}
\arguments{
\item{exp}{exp (#fix).}
}
\description{
check_annotLevels
 
First, check the number of annotations equals the number of columns
in the expression data.
}
\keyword{internal}
