% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bootstrap_args.R
\name{check_bootstrap_args}
\alias{check_bootstrap_args}
\title{check_bootstrap_args}
\usage{
check_bootstrap_args(
  sct_data,
  hits,
  annotLevel,
  reps,
  controlledCT = NULL,
  fix_celltypes = TRUE
)
}
\arguments{
\item{sct_data}{List generated using \link[EWCE]{generate_celltype_data}.}

\item{hits}{List of gene symbols containing the target gene list.
Will automatically be converted to human gene symbols
if \code{geneSizeControl=TRUE}.}

\item{annotLevel}{An integer indicating which level of \code{sct_data} to
analyse (\emph{Default: 1}).}

\item{reps}{Number of random gene lists to generate (\emph{Default: 100},
but should be >=10,000 for publication-quality results).}

\item{controlledCT}{[Optional] If not NULL, and instead is the name of a
cell type, then the bootstrapping controls for expression within that
cell type.}
}
\value{
Null output.
}
\description{
Check the input arguments of the 
\link[EWCE]{bootstrap_enrichment_test}.
}
\keyword{internal}
