% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ewce_expression_data_args.R
\name{check_ewce_expression_data_args}
\alias{check_ewce_expression_data_args}
\title{check_ewce_expression_data_args}
\usage{
check_ewce_expression_data_args(sortBy, tt, thresh)
}
\arguments{
\item{sortBy}{Column name of metric in \code{tt}
which should be used to sort up- from down- regulated genes (Default: "t").}

\item{tt}{Differential expression table.
Can be output of \link[limma]{topTable} function.
Minimum requirement is that one column stores a metric of
increased/decreased expression (i.e. log fold change, t-statistic for
differential expression etc) and another contains gene symbols.}

\item{thresh}{The number of up- and down- regulated genes to be included in
each analysis (Default: 250).}
}
\value{
Null output.
}
\description{
Check the input arguments of the 
\link[EWCE]{ewce_expression_data}.
}
\keyword{internal}
