% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_full_results.R
\name{check_full_results}
\alias{check_full_results}
\title{check_full_results}
\usage{
check_full_results(full_results, sct_data)
}
\arguments{
\item{full_results}{The full output of
\link[EWCE]{bootstrap_enrichment_test} for the same gene list.}

\item{sct_data}{List generated using \link[EWCE]{generate_celltype_data}.}
}
\value{
Null output.
}
\description{
Check full results generated by \link[EWCE]{bootstrap_enrichment_test}.
}
\keyword{internal}
