% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/check_generate_controlled_bootstrapped_geneset.R
\name{check_generate_controlled_bootstrap_geneset}
\alias{check_generate_controlled_bootstrap_geneset}
\title{generate_controlled_bootstrap_geneset}
\usage{
check_generate_controlled_bootstrap_geneset(
  controlledCT,
  annotLevel,
  sct_data,
  hits
)
}
\arguments{
\item{controlledCT}{[Optional] If not NULL, and instead is the name of a
cell type, then the bootstrapping controls for expression within that
cell type.}

\item{annotLevel}{An integer indicating which level of \code{sct_data} to
analyse (\emph{Default: 1}).}

\item{sct_data}{List generated using \link[EWCE]{generate_celltype_data}.}

\item{hits}{List of gene symbols containing the target gene list.
Will automatically be converted to human gene symbols
if \code{geneSizeControl=TRUE}.}
}
\value{
Null output.
}
\description{
Check input arguments to \link[EWCE]{generate_controlled_bootstrap_geneset}.
}
\keyword{internal}
