% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_group_name.R
\name{check_group_name}
\alias{check_group_name}
\title{Check group name}
\usage{
check_group_name(groupName)
}
\arguments{
\item{groupName}{A human readable name for referring to the dataset
being used.}
}
\value{
Null output.
}
\description{
Ensure \code{groupName} argument is provided to 
\link[EWCE]{generate_celltype_data}.
}
\keyword{internal}
