% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nas.R
\name{check_nas}
\alias{check_nas}
\title{Check NAs}
\usage{
check_nas(exp)
}
\arguments{
\item{exp}{Expression matrix.}
}
\value{
Null output.
}
\description{
Check for any NAs in an expression matrix.
}
\keyword{internal}
