% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_numeric}
\alias{check_numeric}
\title{Check numeric}
\usage{
check_numeric(exp)
}
\arguments{
\item{exp}{Input matrix.}
}
\value{
Numeric expression matrix.
}
\description{
Ensure that a matrix is numeric. If not, it will be converted to numeric.
}
\keyword{internal}
