% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sce.R
\name{check_sce}
\alias{check_sce}
\title{Check SingleCellExperiment}
\usage{
check_sce(exp, verbose = TRUE)
}
\value{
List of extracted SCE components.
}
\description{
Check whether \code{exp} is a SingleCellExperiment (SCE) object and extract
the releveant components.
}
\keyword{internal}
