% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene_counts.R
\name{compute_gene_counts}
\alias{compute_gene_counts}
\title{Compute gene counts}
\usage{
compute_gene_counts(bootstrap_list, verbose = TRUE)
}
\arguments{
\item{bootstrap_list}{The output of \code{get_summed_proportions_iterate}.}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table}
}
\description{
Counts the number of times each gene appeared in 
the randomly sampled gene lists.
}
\keyword{internal}
