% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.uninformative.genes.r
\name{construct_SCE}
\alias{construct_SCE}
\title{Construct SingleCellExperiment}
\usage{
construct_SCE(
  exp,
  colData = NULL,
  rowData = NULL,
  save_dir = NULL,
  drop_empty = T,
  replace_HDF5 = F,
  quicksave_HDF5 = F,
  verbose = T
)
}
\description{
Construct SingleCellExperiment
}
\examples{
data("cortex_mrna")
exp <- cortex_mrna$exp
colData <- cortex_mrna$annot
save_dir <- "./cortex_mrna_HDF5Array"
sce <- construct_SCE(exp=exp, colData=colData, save_dir=save_dir)
print(sce)
 @export
 @import DelayedArray
 @import BiocParallel
 @import parallel
 @import SingleCellExperiment
}
