% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_new_ewce_to_old.r
\name{convert_new_ewce_to_old}
\alias{convert_new_ewce_to_old}
\title{convert_new_ewce_to_old}
\usage{
convert_new_ewce_to_old(ctd, lvl)
}
\arguments{
\item{ctd}{A celltype data structure containing $mean_exp and $specificity}

\item{lvl}{The annotation level to extract}
}
\value{
celltype_data The old style data structure
}
\description{
\code{convert_new_ewce_to_old} Used to get an old style EWCE ctd file 
from a new one
}
