% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_old_ewce_to_new.r
\name{convert_old_ewce_to_new}
\alias{convert_old_ewce_to_new}
\title{convert_old_ewce_to_new}
\usage{
convert_old_ewce_to_new(level1, level2)
}
\arguments{
\item{level1}{File path to old level1 of EWCE ctd}

\item{level2}{File path to old level2 of EWCE ctd}
}
\value{
ctd The new style data structure
}
\description{
\code{convert_old_ewce_to_new} Used to get an new style EWCE ctd file (mean_exp/specificity) from old ones (all_scts)
}
