% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.genes.without.1to1.homology.r
\name{convert_orthologues}
\alias{convert_orthologues}
\title{Convert genes from one species to another}
\usage{
convert_orthologues(
  gene_df,
  gene_col = "Gene",
  input_species = "mouse",
  drop_nonhuman_genes = F,
  one_to_one_only = T,
  genes_as_rownames = F,
  verbose = T
)
}
\arguments{
\item{gene_df}{Data.frame containing the genes symbols of the input species.}

\item{gene_col}{Character string indicating either the input species' gene symbols are in a
column with the input species gene symbols (\code{gene_col="<gene_col_name>"})
or are the rownames (\code{gene_col="rownames"}).}

\item{input_species}{Name of the input species ("mouse" or "fly").}

\item{drop_nonhuman_genes}{Drop genes that don't have an orthologue in humans.}

\item{one_to_one_only}{Drop genes that don't have a 1:1 mapping between input species and human
(i.e. drop genes with multiple human orthologues).}

\item{genes_as_rownames}{Whether to return the data.frame with the row names set to the human orthologues.}

\item{verbose}{Show progress messages.}
}
\value{
Data.frame with the gene symbols of the input species ("Gene_orig")
and converted human orthologues ("Gene").
}
\description{
Currently supports: mouse ==> human, fly ==> human.
}
