% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.uninformative.genes.r
\name{convert_to_SCE}
\alias{convert_to_SCE}
\title{Convert expression matrix to SingleCellExperiment}
\usage{
convert_to_SCE(
  exp,
  colData = NULL,
  rowData = NULL,
  save_dir = NULL,
  drop_empty = T,
  replace_HDF5 = F,
  quicksave_HDF5 = F,
  verbose = T
)
}
\description{
Convert expression matrix to SingleCellExperiment
}
\examples{
data("cortex_mrna")
exp <- cortex_mrna$exp
colData <- cortex_mrna$annot
save_dir <- "./cortex_mrna_HDF5Array"
sce <- convert_to_SCE(exp=exp, colData=colData, save_dir=save_dir)
print(sce)
 @export
 @import DelayedArray
 @import BiocParallel
 @import parallel
 @import SingleCellExperiment
}
