% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cortex_mrna}
\alias{cortex_mrna}
\title{Karolinska Cortex/Hippocamus dataset subsample}
\format{An object of class \code{list} of length 2.}
\source{
cortex_mrna  = load.linnarsson.sct.data("expression_mRNA_17-Aug-2014.txt")
cortex_mrna$exp = cortex_mrna$exp[1:200,]
save(cortex_mrna,file="data/cortex_mrna.rda")
}
\usage{
cortex_mrna
}
\description{
Karolinska Cortex/Hippocamus dataset subsample
}
\keyword{datasets}
