% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_DT}
\alias{create_DT}
\title{Make pretty tables}
\usage{
create_DT(DF, caption = "", scrollY = 400)
}
\description{
Take any data.frame and turn it into an interactive, scrollable table
with buttons to download the table in multiple formats.
}
