% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_list_network.R
\name{create_list_network}
\alias{create_list_network}
\title{\code{create_list_network}}
\usage{
create_list_network(data_byGene2, hitGenes_NEW, reps = 10000, no_cores = 1)
}
\value{
List network
}
\description{
Support function for \code{prepare_genesize_control_network}.
}
\keyword{internal}
