% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ctd}
\alias{ctd}
\title{The genes from Karolinska cortex/hippocampus and hypothalamus single cell transcriptome datasets}
\format{An object of class \code{list} of length 2.}
\source{
The datasets were downloaded from the website associated with the paper & GEO and loaded using read_celltype_data PMID:25700174
usethis::use_data(ctd,overwrite = TRUE)
}
\usage{
ctd
}
\description{
All genes from the merged SCT dataset
}
\keyword{datasets}
