% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_to_sce.R
\name{ctd_to_sce}
\alias{ctd_to_sce}
\title{CellTypeDataset to SingleCellExperiment}
\usage{
ctd_to_sce(object, as_sparse = TRUE, as_DelayedArray = TRUE, verbose = TRUE)
}
\value{
SingleCellExperiment
}
\description{
Copied from \href{https://github.com/bschilder/scKirby}{scKirby},
 which is not yet on CRAN or Bioconductor
}
\keyword{internal}
