% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayedarray_normalize.R
\name{delayedarray_normalize}
\alias{delayedarray_normalize}
\title{Efficiently normalize a DelayedArray}
\usage{
delayedarray_normalize(
  exp,
  log_norm = TRUE,
  min_max = TRUE,
  plot_hists = FALSE,
  no_cores = 1
)
}
\arguments{
\item{exp}{Input matrix (e.g. gene expression).}

\item{log_norm}{Whether to first log-normalise \code{exp}
with \link[base]{log1p}.}

\item{min_max}{Whether to min/max-normalise \code{exp}.}

\item{no_cores}{Number of cores to parallelise across.}
}
\description{
The following is a matrix normalization procedure that takes advantage of
functions designed to be more efficient for DelayedArray objects.
}
\keyword{internal}
