% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_nonexpressed_cells.R
\name{drop_nonexpressed_cells}
\alias{drop_nonexpressed_cells}
\title{Drop cells with zero gene counts}
\usage{
drop_nonexpressed_cells(exp, annotLevels, verbose = TRUE)
}
\arguments{
\item{exp}{Gene expression matrix.}

\item{annotLevels}{Cell-wise annotations to be subset
if some cells are dropped.}

\item{verbose}{Print messages.}
}
\description{
Remove columns (cells) in which (gene) counts sum to zero.
}
\keyword{internal}
