% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_nonexpressed_genes.R
\name{drop_nonexpressed_genes}
\alias{drop_nonexpressed_genes}
\title{Drop genes with zero counts}
\usage{
drop_nonexpressed_genes(exp, verbose = TRUE)
}
\arguments{
\item{exp}{Gene expression matrix.}

\item{verbose}{Print messages.}
}
\value{
List of filtered \code{exp}.
}
\description{
Remove rows (genes) in which counts sum to zero.
}
\keyword{internal}
