% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_df.R
\name{dt_to_df}
\alias{dt_to_df}
\title{Convert a \code{data.table} to a \code{data.frame}.}
\usage{
dt_to_df(exp)
}
\description{
Converts a \code{data.table} to a \code{data.frame} by setting the
 first column as the rownames.
}
\keyword{internal}
