% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{EWCE-package}
\alias{EWCE}
\alias{EWCE-package}
\title{EWCE: Expression Weighted Celltype Enrichment}
\description{
Used to determine which cell types are enriched within gene lists. The package provides tools for testing enrichments within simple gene lists (such as human disease associated genes) and those resulting from differential expression studies. The package does not depend upon any particular Single Cell Transcriptome dataset and user defined datasets can be loaded in and used in the analyses.
}
\details{
EWCE: Expression Weighted Celltype Enrichment

Used to determine which cell types are enriched within gene lists.
The package provides tools for testing enrichments within simple gene lists
 (such as human disease associated genes) and those resulting from
 differential expression studies.

The package does not depend upon any particular Single Cell Transcriptome
 dataset and user defined datasets can be loaded in and used in the analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NathanSkene/EWCE}
  \item Report bugs at \url{https://github.com/NathanSkene/EWCE/issues}
}

}
\author{
\strong{Maintainer}: Alan Murphy \email{alanmurph94@hotmail.com} (\href{https://orcid.org/0000-0002-2487-8753}{ORCID})

Authors:
\itemize{
  \item Brian Schilder \email{brian_schilder@alumni.brown.edu} (\href{https://orcid.org/0000-0001-5949-2191}{ORCID})
  \item Nathan Skene \email{nathan.skene@gmail.com} (\href{https://orcid.org/0000-0002-6807-3180}{ORCID})
}

}
