% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_genes_without_1to1_homology.r
\name{filter_genes_without_1to1_homolog}
\alias{filter_genes_without_1to1_homolog}
\title{filter_genes_without_1to1_homolog}
\usage{
filter_genes_without_1to1_homolog(filenames)
}
\arguments{
\item{filenames}{Array of filenames for sct_data saved as .Rda files}
}
\value{
Array of filenames included the ones with only 1:1 homologs
}
\description{
\code{filter_genes_without_1to1_homolog} Takes the filenames of
celltype_data files, loads them, and drops any genes
which don't have a 1:1 homolog based on biomart. The new files are saved to
disc, appending '_1to1only' to the file
tag of the previous file.
}
\examples{
library(ewceData)
# Load the single cell data
cortex_mrna <- cortex_mrna()
expData <- cortex_mrna$exp
expData <- expData[1:100, ] # Use only a subset to keep the example quick
l1 <- cortex_mrna$annot$level1class
l2 <- cortex_mrna$annot$level2class
annotLevels <- list(level1class = l1, level2class = l2)
fNames_ALLCELLS <- generate_celltype_data(exp = expData,
    annotLevels = annotLevels,
    groupName = "allKImouse")
fNames_ALLCELLS <- filter_genes_without_1to1_homolog(fNames_ALLCELLS)
}
