% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_nonorthologs.R
\name{filter_nonorthologs}
\alias{filter_nonorthologs}
\title{Filter non-orthologs}
\usage{
filter_nonorthologs(
  filenames,
  input_species = NULL,
  convert_nonhuman_genes = TRUE,
  annot_levels = NULL,
  suffix = "_orthologs",
  method = "homologene",
  verbose = TRUE
)
}
\arguments{
\item{filenames}{List of file names for sct_data saved as \emph{.rda} files.}

\item{input_species}{Which species the gene names in \code{exp} come from.}

\item{convert_nonhuman_genes}{Whether to convert the \code{exp}
row names to human gene names.}

\item{annot_levels}{[Optional] Names of each annotation level.}

\item{suffix}{Suffix to add to the file name (right before \emph{.rda}).}

\item{method}{R package to to use for gene mapping:
\itemize{
\item{\code{"gprofiler"} : Slower but more species and genes.}
\item{\code{"homologene"} : Faster but fewer species and genes.}
\item{\code{"babelgene"} : Faster but fewer species and genes.
Also gives consensus scores for each gene mapping based on a
 several different data sources.}
}}

\item{verbose}{Print messages.}
}
\value{
List of the filtered CellTypeData file names.
}
\description{
\code{filter_nonorthologs} Takes the filenames of CellTypeData files,
loads them,  drops any genes which don't have a 1:1 orthologs with humans,
and then convert the gene to human orthologs.
The new files are then saved to disk, appending
'_orthologs' to the file name.
}
\details{
\bold{Note:} This function replaces the original
 \code{filter_genes_without_1to1_homolog} function.
\code{filter_genes_without_1to1_homolog} is
now a wrapper for \code{filter_nonorthologs}.
}
\examples{
# Load the single cell data
ctd <- ewceData::ctd()
tmp <- tempfile()
save(ctd, file = tmp)
fNames_ALLCELLS_orths <- EWCE::filter_nonorthologs(filenames = tmp)
}
