% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.genes.without.1to1.homology.r
\name{filter_nonorthologues}
\alias{filter_nonorthologues}
\title{filter_nonorthologues}
\usage{
filter_nonorthologues(
  filenames,
  input_species = "mouse",
  annot_levels = NULL,
  suffix = "_orthologues",
  verbose = T
)
}
\arguments{
\item{filenames}{List of file names for sct_data saved as \emph{.rda} files.}

\item{input_species}{Name of the input species ("mouse" or "fly").}

\item{suffix}{Suffix to add to the file name (right before \emph{.rda}).}

\item{verbose}{Show progress messages.}
}
\value{
List of the filtered CellTypeData file names.
}
\description{
\code{filter_nonorthologues} Takes the filenames of CellTypeData files, loads them, and drops any genes
which don't have a 1:1 orthologues based on \pkg{homologene}.
The new files are saved to disk, appending '_orthologues' to the file name.
}
\examples{
# Load the single cell data
data(cortex_mrna)
expData = cortex_mrna$exp
expData = expData[1:500,] # Use only a subset to keep the example quick
l1=cortex_mrna$annot$level1class
l2=cortex_mrna$annot$level2class
annotLevels = list(level1class=l1,level2class=l2)
fNames_ALLCELLS = generate.celltype.data(exp=expData,annotLevels=annotLevels,groupName="allKImouse")
fNames_ALLCELLS_orths = filter_nonorthologues(fNames_ALLCELLS)
}
