% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_celltype_markers.R
\name{find_celltype_markers}
\alias{find_celltype_markers}
\title{Find the top gene markers for each cell-type}
\usage{
find_celltype_markers(ctd, level = 2)
}
\description{
Find the top gene markers for each cell-type
}
