% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.bad.mgi.symbols.r
\name{fix.bad.mgi.symbols}
\alias{fix.bad.mgi.symbols}
\title{fix.bad.mgi.symbols
- Given an expression matrix, wherein the rows are supposed to be MGI 
symbols, find those symbols which are not official MGI symbols, then
check in the MGI synonm database for whether they match to a proper MGI 
symbol. Where a symbol is found to be an aliases for a gene that is already
in the dataset, the combined reads are summed together.}
\usage{
fix.bad.mgi.symbols(exp, mrk_file_path = NULL, printAllBadSymbols = FALSE)
}
\arguments{
\item{exp}{An expression matrix where the rows are MGI symbols}

\item{mrk_file_path}{Path to the MRK_List2 file which can be downloaded 
from www.informatics.jax.org/downloads/reports/index.html}

\item{printAllBadSymbols}{Output to console all the bad gene symbols}
}
\value{
Returns the expression matrix with the rownames corrected and rows 
representing the same gene merged. If no corrections are necessary, input 
expression matrix is returned.
}
\description{
Also checks whether any gene names contain "Sep", "Mar" or "Feb". 
These should be checked for any suggestion that excel has corrupted the 
gene names.
}
\examples{
# Load the single cell data
data("cortex_mrna")
cortex_mrna$exp <- fix.bad.mgi.symbols(cortex_mrna$exp)
}
