% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_bad_hgnc_symbols.r
\name{fix_bad_hgnc_symbols}
\alias{fix_bad_hgnc_symbols}
\title{fix_bad_hgnc_symbols
- Given an expression matrix, wherein the rows are supposed to be HGNC
symbols, find those symbols which are not official HGNC symbols, then
correct them if possible. Return the expression matrix with corrected
symbols.}
\usage{
fix_bad_hgnc_symbols(exp, dropNonHGNC = FALSE)
}
\arguments{
\item{exp}{An expression matrix where the rows are HGNC symbols or an SingleCellExperiment (SCE) or other Ranged Summarized Experiment (SE) type object.}

\item{dropNonHGNC}{Boolean. Should symbols not recognised as HGNC symbols
be dropped?}
}
\value{
Returns the expression matrix with the rownames corrected and rows
representing the same gene merged. If a SingleCellExperiment (SCE) or other
Ranged Summarized Experiment (SE) type object was inputted this will be
returned with the corrected expression matrix under counts.
}
\description{
fix_bad_hgnc_symbols
- Given an expression matrix, wherein the rows are supposed to be HGNC
symbols, find those symbols which are not official HGNC symbols, then
correct them if possible. Return the expression matrix with corrected
symbols.
}
