% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_celltype_names_full_results.R
\name{fix_celltype_names_full_results}
\alias{fix_celltype_names_full_results}
\title{Fix celltype name in full results}
\usage{
fix_celltype_names_full_results(full_results, verbose = TRUE)
}
\arguments{
\item{full_results}{Cell-type enrichment results generated by
\link[EWCE]{bootstrap_enrichment_test}.}

\item{verbose}{Print messages.}
}
\value{
Fixed full results.
}
\description{
Aligns celltype names in full results generated by
\link[EWCE]{bootstrap_enrichment_test} with the standardised
CellTypeDataset (CTD) produced by \link[EWCE]{standardise_ctd}.
}
\keyword{internal}
