% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{full_results}
\alias{full_results}
\title{Schizophrenia susceptibility genes from CLOZUK}
\format{
An object of class \code{list} of length 3.
}
\source{
Extended data tables downloaded from: http://www.biorxiv.org/content/early/2016/08/09/068593.figures-only
Ran the relevant parts of the vignette
usethis::use_data(full_results)
}
\usage{
full_results
}
\description{
Schizophrenia susceptibility genes from CLOZUK
}
\keyword{datasets}
