% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.celltype.data.r
\name{generate.celltype.data}
\alias{generate.celltype.data}
\title{generate.celltype.data}
\usage{
generate.celltype.data(exp, level1class, level2class, groupName, thresh = 0,
  trim = 0)
}
\arguments{
\item{exp}{Numerical matrix with row for each gene and column for each cell. Row names are MGI/HGNC gene symbols. Column names are cell IDs which can be cross referenced against the annot data frame.}

\item{level1class}{Array of strings containing the level 1 cell type names associated with each column in exp}

\item{level2class}{Array of strings containing the level 2 cell type names associated with each column in exp}

\item{groupName}{A human readable name for refering to the dataset being loaded}

\item{thresh}{Expression threshold value below which a gene is dropped (default=0). We do not recommend changing from the default.}

\item{trim}{Value determining how the trimmed mean is calculated (range: 0 to 0.5, default=0). We do not recommend changing from the default.}
}
\value{
Filenames for the saved celltype_data files
}
\description{
\code{generate.celltype.data} Takes expression & cell type annotations and creates celltype_data files which contain the mean and specificity matrices
}
\examples{
# Load the single cell data
data("cortex_mrna")
expData = cortex_mrna$exp
l1=cortex_mrna$annot$level1class
l2=cortex_mrna$annot$level2class
fNames_ALLCELLS = generate.celltype.data(exp=expData,l1,l2,"allKImouse",thresh=0,trim=0)
}
