% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_controlled_bootstrap_geneset.r
\name{generate_controlled_bootstrap_geneset}
\alias{generate_controlled_bootstrap_geneset}
\title{generate_controlled_bootstrap_geneset}
\usage{
generate_controlled_bootstrap_geneset(
  hitGenes,
  sct_data,
  combinedGenes,
  annotLevel,
  reps,
  controlledCT = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{hitGenes}{Array of gene names. The target gene set.}

\item{sct_data}{The cell type data list (with specificity and mean_exp)}

\item{annotLevel}{The level of annotation in sct_data to analyse}

\item{reps}{The number of gene lists to sample}

\item{controlledCT}{Name of a cell type (from colnames of
\code{sct_data[[x]]$specificity}).}

\item{verbose}{Print messages.}
}
\value{
Matrix of genes
 (such that \code{nrows=length(hitGenes)} and \code{ncols=reps}), where each
column is a gene list.
}
\description{
Used to generated cell type-controlled bootstrapped gene sets.
}
\details{
See \link[EWCE]{controlled_geneset_enrichment} for examples.
}
\keyword{internal}
