% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.celltype.table.r
\name{get.celltype.table}
\alias{get.celltype.table}
\title{get.celltype.table}
\usage{
get.celltype.table(annot)
}
\arguments{
\item{annot}{An annotation dataframe, which columns named 'level1class', 
'level2class' and 'dataset_name'}
}
\value{
A dataframe with columns 'name', 'level', 'freq' and 'dataset_name'
}
\description{
\code{get.celltype.table} Generates a table that can be used for 
supplemenary tables of publications.
The table lists how many cells are associated with each celltype, the level 
of annotation, and the dataset from which it was generated.
}
\examples{
library(ewceData)
# See PrepLDSC.Rmd for origin of merged_ALLCELLS$annot
cortex_mrna <- cortex_mrna()
cortex_mrna$annot$dataset_name <- "cortex_mrna"
celltype_table <- get.celltype.table(cortex_mrna$annot)
}
