% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.length.of.longest.string.r
\name{get.length.of.longest.string}
\alias{get.length.of.longest.string}
\title{get.length.of.longest.string}
\usage{
get.length.of.longest.string(x)
}
\arguments{
\item{x}{Array of names}
}
\value{
Length of the longest name [integer]
}
\description{
\code{get.length.of.longest.string} Given an array of strings (i.e. cellnames). Find the length of the longest
in terms of characters. This can be used to determine the width of PDF figures.
}
\examples{
get.length.of.longest.string(c("nathan","tom"))
}
