% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctd_levels.R
\name{get_ctd_levels}
\alias{get_ctd_levels}
\title{Get the names of CellTypeDataset levels}
\usage{
get_ctd_levels(ctd, max_only = FALSE)
}
\arguments{
\item{ctd}{CellTypeDataset.}

\item{max_only}{Only return the level with the greatest depth
(e.g. \code{"level3"} in \code{c("level1","level2","level3")}).}
}
\description{
Returns the level names of a CellTypeDataset. If none are available,
will instead return a vector of numbers (one number per level).
}
