% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctd_matrix_names.R
\name{get_ctd_matrix_names}
\alias{get_ctd_matrix_names}
\title{Get CTD matrix names}
\usage{
get_ctd_matrix_names(
  ctd = NULL,
  matrices = c("mean_exp", "median_exp", "specificity", "median_specificity",
    "specificity_quantiles"),
  verbose = TRUE
)
}
\arguments{
\item{ctd}{CellTypeDataset. If set to \code{NULL} (default), 
will simply return all possible matrix names.}

\item{matrices}{Matrix names to search for.}

\item{verbose}{Print messages.}
}
\value{
List of matrix names.
}
\description{
Find the names of all data matrices in a CellTypeDataset.
}
\keyword{internal}
