% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_theme.R
\name{get_graph_theme}
\alias{get_graph_theme}
\title{Get graph theme}
\usage{
get_graph_theme()
}
\value{
\code{ggplot2} graph theme.
}
\description{
Get graph theme for plots created by 
\link[EWCE]{generate_bootstrap_plots_for_transcriptome}.
}
\keyword{internal}
