% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_results.R
\name{get_sig_results}
\alias{get_sig_results}
\title{Extract significant results}
\usage{
get_sig_results(
  full_results,
  mtc_method = "BH",
  q_threshold = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{full_results}{Output of \link[EWCE]{bootstrap_enrichment_test}.}

\item{mtc_method}{Multiple-testing correction method
(passed to \link[stats]{p.adjust}).}

\item{q_threshold}{Maximum multiple-testing-corrected p-value to include.}

\item{verbose}{Print messages.}
}
\value{
Filtered enrichment results table.
}
\description{
Extract significant results from output of
\link[EWCE]{bootstrap_enrichment_test}.
}
\keyword{internal}
