% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_celltypedataset.R
\name{is_celltypedataset}
\alias{is_celltypedataset}
\title{Check whether object is a CellTypeDataset}
\usage{
is_celltypedataset(ctd)
}
\arguments{
\item{ctd}{Object.}
}
\value{
boolean
}
\description{
Check whether an object is a CellTypeDataset.
}
\keyword{internal}
