% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ctd_standardised.R
\name{is_ctd_standardised}
\alias{is_ctd_standardised}
\title{Check whether a CellTypeDataset is standardised}
\usage{
is_ctd_standardised(ctd)
}
\arguments{
\item{ctd}{CellTypeDataset.}
}
\value{
Whether the \code{ctd} is standardised.
}
\description{
Check whether a CellTypeDataset was previously standardised
using \link[EWCE]{standardise_ctd}.
}
\keyword{internal}
