% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_delayed_array.R
\name{is_delayed_array}
\alias{is_delayed_array}
\title{Assess whether an object is a DelayedArray.}
\usage{
is_delayed_array(X)
}
\arguments{
\item{X}{Object.}
}
\description{
Assess whether an object is a DelayedArray or one of
 its derived object types.
}
