% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_matrix.R
\name{is_matrix}
\alias{is_matrix}
\title{Assess whether an object is a Matrix}
\usage{
is_matrix(X)
}
\arguments{
\item{X}{Object.}
}
\value{
boolean
}
\description{
Assess whether an object is a Matrix or one of
 its derived object types.
}
