% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sparse_matrix.R
\name{is_sparse_matrix}
\alias{is_sparse_matrix}
\title{Assess whether an object is a sparse matrix}
\usage{
is_sparse_matrix(X)
}
\arguments{
\item{X}{Object.}
}
\value{
boolean
}
\description{
Assess whether an object is a sparse matrix or one of
 its derived object types.
}
