% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_species.R
\name{list_species}
\alias{list_species}
\title{List all species}
\usage{
list_species(verbose = TRUE)
}
\arguments{
\item{verbose}{Print messages.}
}
\value{
List of species EWCE can input/output genes as.
}
\description{
List all species that EWCE can convert genes from/to.
Wrapper function for \link[orthogene]{map_species}.
}
\examples{
list_species()
}
