% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_ctd_depth.R
\name{max_ctd_depth}
\alias{max_ctd_depth}
\title{Get max CTD depth}
\usage{
max_ctd_depth(CTD_list)
}
\arguments{
\item{CTD_list}{A list of CellTypeDataset objects.}
}
\value{
integer
}
\description{
Get the maximum level depth from a list of CellTypeDataset objects.
}
\keyword{internal}
