% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce_list.R
\name{merge_sce_list}
\alias{merge_sce_list}
\title{Merge of list of SingleCellExperiment objects}
\usage{
merge_sce_list(
  SCE_lists = NULL,
  parent_folder = NULL,
  pattern = ".rds$",
  merge_levels = seq(1, 5),
  gene_union = TRUE,
  as_sparse = TRUE,
  as_DelayedArray = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{SCE_lists}{A list of \link[SingleCellExperiment]{SingleCellExperiment}
objects.}

\item{parent_folder}{Can supply the path to a folder
instead of \code{SCE_lists}. Any \link[SingleCellExperiment]{SingleCellExperiment}
 objects matching \code{pattern} will be imported.}

\item{merge_levels}{CellTypeDataset levels to merge}
}
\description{
Merge of list of CellTypeDatasets stored as
\link[SingleCellExperiment]{SingleCellExperiment} objects
into one \link[SingleCellExperiment]{SingleCellExperiment} object.
}
\keyword{internal}
