% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_parallel.R
\name{message_parallel}
\alias{message_parallel}
\title{Print messages}
\usage{
message_parallel(...)
}
\arguments{
\item{...}{Message input.}
}
\description{
Print messages even from within parallelised functions.
}
\keyword{internal}
