% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mgi_synonym_data}
\alias{mgi_synonym_data}
\title{MGI synonym data}
\format{
An object of class \code{data.frame} with 55834 rows and 12 columns.
}
\source{
link2 <- "http://www.informatics.jax.org/downloads/reports/MRK_List2.rpt"
download.file(link2, destfile="MRK_List2.rpt")
mgi_synonym_data = read.csv(mrk_file_path,sep="\\t",stringsAsFactors = FALSE)
mgi_synonym_data = mgi_data[!mgi_data$Marker.Synonyms..pipe.separated.=="",]
save(mgi_synonym_data,file="data/mgi_synonym_data.rda", compress='xz')
}
\usage{
mgi_synonym_data
}
\description{
MGI synonym data
}
\keyword{datasets}
