% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myScalesComma.R
\name{myScalesComma}
\alias{myScalesComma}
\title{\code{myScalesComma}}
\usage{
myScalesComma(x)
}
\description{
Adjusts \pkg{ggplot2} label display. See \link[scales]{comma} for details.
Support function for \link[EWCE]{plot_log_bootstrap_distributions}.
}
\keyword{internal}
