% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{A small example version of the PBMC dataset (Seurat format)}
\format{
An object of class \code{Seurat} with 230 rows and 80 columns.
}
\source{
library(Seurat)
data(pbmc_small)
pbmc_small <- Seurat::UpdateSeuratObject(pbmc_small)
usethis::use_data(pbmc_small, internal = F, overwrite = T)
}
\usage{
pbmc_small
}
\description{
A subsetted version of 10X Genomics' 3k PBMC dataset
}
\references{
\href{Seurat::pbmc_small}{https://rdrr.io/cran/Seurat/man/pbmc_small.html}.
}
\keyword{datasets}
