% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrap_plots.R
\name{plot_bootstrap_plots}
\alias{plot_bootstrap_plots}
\title{Bootstrap plot}
\usage{
plot_bootstrap_plots(
  dat,
  tag,
  listFileName,
  cc,
  showGNameThresh,
  graph_theme,
  maxX,
  save_dir
)
}
\value{
Null result.
}
\description{
Plot results of \link[EWCE]{generate_bootstrap_plots_for_transcriptome}.
}
\keyword{internal}
