% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_metrics.R
\name{plot_gene_metrics}
\alias{plot_gene_metrics}
\title{plot_gene_metrics}
\usage{
plot_gene_metrics(
  ctd,
  level = 2,
  metric = "specificity",
  n_genes = NULL,
  title = metric,
  interactive = F,
  remove_numeric_groups = F,
  show_plot = T
)
}
\arguments{
\item{ctd}{CellTypeData object produced by \pkg{EWCE}.}

\item{level}{Which annotation level to plot.}

\item{metric}{"mean_exp", "specificity", or "specificity_quantiles".}

\item{n_genes}{Limit plot to just the top N genes (according to \code{metric}).}

\item{title}{Plot title.}

\item{interactive}{Make the plot interactive with \pkg{plotly}.}

\item{remove_numeric_groups}{Easy way to remove unlabeled levels (i.e. clusters only identified by numbers).}
}
\description{
Convenience functiont to quickly plot multiple aspects of a CellTypeData (CTD) object.
}
\examples{
data("ctd")
gp <- plot_gene_metrics(ctd = ctd, metric = "specificity", level = 2, n_genes = 10)
}
