% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log_bootstrap_distributions.R
\name{plot_log_bootstrap_distributions}
\alias{plot_log_bootstrap_distributions}
\title{Plot log bootstrap distributions}
\usage{
plot_log_bootstrap_distributions(
  dat,
  exp_mats,
  cc,
  hit_exp,
  tag,
  listFileName,
  graph_theme,
  save_dir
)
}
\value{
Null result.
}
\description{
Plot results of \link[EWCE]{generate_bootstrap_plots_for_transcriptome}.
}
\keyword{internal}
