% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.dendro.r
\name{prep.dendro}
\alias{prep.dendro}
\title{prep.dendro}
\usage{
prep.dendro(ctdIN)
}
\arguments{
\item{ctdIN}{A single annotLevel of a ctd, i.e. ctd[[1]] (the function is 
intended to be used via apply)}
}
\value{
A ctd with dendrogram plotting info added
}
\description{
\code{prep.dendro} adds a dendrogram to a cts
}
\examples{
ctd <- lapply(ctd, EWCE::bin.specificity.into.quantiles, numberOfBins = 40)
ctd <- lapply(ctd, EWCE::prep.dendro)
}
