% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dge.R
\name{report_dge}
\alias{report_dge}
\title{Report DGE}
\usage{
report_dge(exp, keep_genes, adj_pval_thresh = 0.05, verbose = TRUE)
}
\arguments{
\item{exp}{Gene expression matrix.}

\item{keep_genes}{Genes kept after DGE.}

\item{adj_pval_thresh}{Minimum differential expression significance
that a gene must demonstrate across \code{level2annot} (i.e. cell types).}

\item{verbose}{Print messages.
#' @inheritParams orthogene::convert_orthologs}
}
\value{
Null output.
}
\description{
Report differential gene expression (DGE) results
}
\keyword{internal}
