% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_results.R
\name{report_results}
\alias{report_results}
\title{Report cell type enrichment results}
\usage{
report_results(results, sig_thresh = 0.05, verbose = TRUE)
}
\value{
NULL output.
}
\description{
Report cell type enrichment results generated by
\link[EWCE]{bootstrap_enrichment_test}.
}
\keyword{internal}
