% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.uninformative.genes.r
\name{run_glmGamPoi_DE}
\alias{run_glmGamPoi_DE}
\title{Run Gamma-Poisson Generalized Linear regression (glmGamPoi)}
\usage{
run_glmGamPoi_DE(
  sce,
  level2annot,
  pseudobulk_by = NULL,
  pval_adjust_method = "BH",
  adj_pval_thresh = 1e-05,
  on_disk = T,
  return_as_SCE = T,
  verbose = T
)
}
\description{
Installation troubleshooting:
1. Try to install using  Bioconductor: \code{BiocManager::install("glmGamPoi")}
2. If that fails, try installing directly from GitHub: \code{devtools::install_github("const-ae/glmGamPoi")}
3. If that fails due to an error like \code{ld: warning: directory not found for option '-L/usr/local/gfortran/lib/gcc/x86_64-apple-darwin15/6.1.0'}
follow \href{https://stackoverflow.com/questions/35999874/mac-os-x-r-error-ld-warning-directory-not-found-for-option}{these instructions} to fix this.
}
\examples{
data("cortex_mrna")
exp <- cortex_mrna$exp[1:300,]
colData <- cortex_mrna$annot
save_dir <- "~/Desktop/cortex_mrna_HDF5Array"
sce <- convert_to_SCE(exp=exp, colData=colData, save_dir=save_dir, replace_HDF5 = T)
exp <- run_glmGamPoi(sce, level2annot="level2class")
}
