% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mast.r
\name{run_mast}
\alias{run_mast}
\title{Run DGE:  \pkg{MAST}}
\usage{
run_mast(exp, level2annot, no_cores = 1, ...)
}
\arguments{
\item{exp}{Expression matrix with gene names as rownames.}

\item{level2annot}{Array of cell types, with each sequentially corresponding
a column in the expression matrix}

\item{no_cores}{Number of cores to parallelise across.
Set to \code{NULL} to automatically optimise.}

\item{...}{Additional arguments to be passed to the selected DGE method.}
}
\value{
\code{MAST} results
}
\description{
Run Differential Gene Expression with \pkg{MAST}.
}
\keyword{internal}
