% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_filepath.R
\name{sce_filepath}
\alias{sce_filepath}
\title{Construct SingleCellExperiment file path}
\usage{
sce_filepath(sce, sce_save_dir = NULL)
}
\value{
SingleCellExperiment file path.
}
\description{
Construct a save path for a SingleCellExperiment (SCE).
}
\keyword{internal}
