% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_lists_apply.R
\name{sce_lists_apply}
\alias{sce_lists_apply}
\title{sce_lists_apply}
\usage{
sce_lists_apply(
  SCE_lists,
  return_genes = FALSE,
  level = 2,
  as_matrix = FALSE,
  as_DelayedArray = FALSE
)
}
\description{
Support function for \code{EWCE::merge_sce_list}.
}
\keyword{internal}
