% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce_merged_apply.R
\name{sce_merged_apply}
\alias{sce_merged_apply}
\title{sce_merged_apply}
\usage{
sce_merged_apply(SCE_merged, as_sparse = TRUE, as_DelayedArray = FALSE)
}
\value{
Merged SingleCellExperiment.
}
\description{
Merge a list of SingleCellExperiments.
}
\keyword{internal}
