% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_graph.R
\name{theme_graph}
\alias{theme_graph}
\title{Get graph theme}
\usage{
theme_graph()
}
\value{
\code{ggplot2} graph theme.
}
\description{
Get graph theme for plots created by 
\link[EWCE]{generate_bootstrap_plots_for_transcriptome}.
}
\keyword{internal}
