% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_dataframe.R
\name{to_dataframe}
\alias{to_dataframe}
\title{Convert object to data.frame}
\usage{
to_dataframe(X, verbose = TRUE)
}
\arguments{
\item{X}{Object.}

\item{verbose}{Print messages.}
}
\description{
Convert a variety of object types to data.frame format.
}
\keyword{internal}
