% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sparse_matrix.R
\name{to_sparse_matrix}
\alias{to_sparse_matrix}
\title{Convert object to sparse matrix}
\usage{
to_sparse_matrix(exp, as_sparse = TRUE, verbose = TRUE)
}
\arguments{
\item{exp}{Object.}

\item{as_sparse}{Whether to convert \code{exp} to sparse matrix}

\item{verbose}{Print messages.}
}
\value{
Sparse matrix.
}
\description{
Convert a variety of object types to sparse matrix format.
}
\keyword{internal}
