% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_plot.R
\name{bootstrap_plot}
\alias{bootstrap_plot}
\title{Bootstrap plot}
\usage{
bootstrap_plot(
  gene_data,
  exp_mats = NULL,
  save_dir = file.path(tempdir(), "BootstrapPlots"),
  listFileName,
  signif_ct = NULL,
  hit_thresh = 25,
  facets = "CellType",
  scales = "free_x",
  show_plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gene_data}{Output from \link[EWCE]{compute_gene_scores}.}

\item{exp_mats}{Output of \code{generate_bootstrap_plots_exp_mats}.}

\item{save_dir}{Directory to save plots to.}

\item{listFileName}{listFileName}

\item{signif_ct}{Significant celltypes to include the plots.}

\item{facets}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{rows}
and \code{cols} instead.}

\item{scales}{Are scales shared across all facets (the default,
\code{"fixed"}), or do they vary across rows (\code{"free_x"}),
columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{show_plot}{Print the plot.}
}
\value{
Null output.
}
\description{
Plot bootstrap enrichment results. 
Support function for \link[EWCE]{generate_bootstrap_plots}.
}
\keyword{internal}
