% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_quantiles.R
\name{calc_quantiles}
\alias{calc_quantiles}
\title{Calculate quantiles}
\usage{
calc_quantiles(v, n_quantiles = 10, report_filters = TRUE)
}
\value{
Quantiles.
}
\description{
Calculate quantiles
}
\keyword{internal}
