% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ewce_genelist_input.r
\name{check_ewce_genelist_inputs}
\alias{check_ewce_genelist_inputs}
\title{check_ewce_genelist_inputs}
\usage{
check_ewce_genelist_inputs(
  sct_data,
  hits,
  bg = NULL,
  genelistSpecies = NULL,
  sctSpecies = NULL,
  output_species = "human",
  geneSizeControl = FALSE,
  standardise = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{sct_data}{List generated using \link[EWCE]{generate_celltype_data}.}

\item{hits}{List of gene symbols containing the target gene list.
Will automatically be converted to human gene symbols
if \code{geneSizeControl=TRUE}.}

\item{bg}{List of gene symbols containing the background gene list
(including hit genes). If \code{bg=NULL},
 an appropriate gene background will be created automatically.
if \code{geneSizeControl=TRUE}.}

\item{genelistSpecies}{Species that \code{hits} genes came from
(no longer limited to just "mouse" and "human").}

\item{sctSpecies}{Species that \code{sct_data} came from
(no longer limited to just "mouse" and "human").}

\item{output_species}{Species to convert \code{sct_data} and \code{hits} to
(Default: "human").}

\item{geneSizeControl}{Whether you want to control for
GC content and transcript length. Recommended if the gene list originates
from genetic studies (\emph{Default: FALSE}).
 If set to \code{TRUE}, then \code{hits} must be from humans.
should be used rather than mouse.}

\item{standardise}{If \code{input_species==output_species},
should the genes be standardised using \link[orthogene]{map_genes}?}

\item{verbose}{Print messages.}
}
\value{
A list containing
\itemize{
  \item \code{hits}: Array of MGI/HGNC gene symbols containing the target
  gene list.
  \item \code{bg}: Array of MGI/HGNC gene symbols containing the background
  gene list.
}
}
\description{
\code{check_ewce_genelist_inputs} Is used to check that hits and bg gene
lists passed to EWCE are setup correctly. Checks they are the
appropriate length.
Checks all hits genes are in bg. Checks the species match and if not
reduces to 1:1 orthologs.
}
\examples{
ctd <- ewceData::ctd()
example_genelist <- ewceData::example_genelist()

# Called from "bootstrap_enrichment_test()" and "generate_bootstrap_plots()"
checkedLists <- EWCE::check_ewce_genelist_inputs(
    sct_data = ctd,
    hits = example_genelist,
    sctSpecies = "mouse",
    genelistSpecies = "human"
)
}
