% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_species.R
\name{check_species}
\alias{check_species}
\title{Check species}
\usage{
check_species(genelistSpecies = NULL, sctSpecies = NULL, verbose = TRUE)
}
\arguments{
\item{genelistSpecies}{Species that \code{hits} genes came from
(no longer limited to just "mouse" and "human").
See \link[EWCE]{list_species} for all available species.}

\item{sctSpecies}{Species that \code{sct_data} came from
(no longer limited to just "mouse" and "human").
See \link[EWCE]{list_species} for all available species.}

\item{verbose}{Print messages.}
}
\value{
List of corrected species names.
}
\description{
If species arguments are \code{NULL}, set default species.
}
\keyword{internal}
