% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cortex_mrna}
\alias{cortex_mrna}
\title{Karolinska Cortex/Hippocamus dataset subsample}
\format{An object of class \code{list} of length 2.}
\source{
download.file("goo.gl/r5Y24y",destfile="expression_mRNA_17-Aug-2014.txt")
path = "expression_mRNA_17-Aug-2014.txt"
cortex_mrna  = load.linnarsson.sct.data("expression_mRNA_17-Aug-2014.txt")
save(cortex_mrna,file="data/cortex_mrna.rda")
}
\usage{
cortex_mrna
}
\description{
Karolinska Cortex/Hippocamus dataset subsample
}
\keyword{datasets}
