% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_to_sce.R
\name{ctd_to_sce}
\alias{ctd_to_sce}
\title{CellTypeDataset to SingleCellExperiment}
\usage{
ctd_to_sce(object, as_sparse = TRUE, as_DelayedArray = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{CellTypeDataset object.}

\item{as_sparse}{Store SingleCellExperiment matrices as sparse.}

\item{as_DelayedArray}{Store SingleCellExperiment matrices as DelayedArray.}

\item{verbose}{Print messages.}
}
\value{
SingleCellExperiment
}
\description{
Copied from \href{https://github.com/neurogenomics/scKirby}{scKirby},
 which is not yet on CRAN or Bioconductor.
}
\examples{
ctd <- ewceData::ctd()
sce <- EWCE::ctd_to_sce(ctd)
}
