% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.uninformative.genes.r
\name{drop.uninformative.genes}
\alias{drop.uninformative.genes}
\title{drop.uninformative.genes}
\source{
#' \href{https://petehaitch.github.io/BioC2020_DelayedArray_workshop/articles/Effectively_using_the_DelayedArray_framework_for_users.html}{DelayedArray workshop}
}
\usage{
drop.uninformative.genes(
  exp,
  level2annot,
  DGE_method = "limma",
  min_variance_decile = NULL,
  adj_pval_thresh = 1e-05,
  drop_nonhuman_genes = F,
  input_species = NULL,
  sce_save_dir = NULL,
  return_sce = F,
  verbose = T,
  ...
)
}
\arguments{
\item{exp}{Expression matrix with gene names as rownames.}

\item{DGE_method}{Which method to use for the Differential Gene Expression (DGE) step.}

\item{min_variance_decile}{If `min_variance_decile!=NULL`, calculates the variance of the mean gene expression  across `level2annot` (i.e. cell-types),
and then removes any genes that are below `min_variance_decile` (on a 0-1 scale).}

\item{adj_pval_thresh}{Minimum differential expression significance
that a gene must demonstrate across `level2annot` (i.e. cell-types).}

\item{drop_nonhuman_genes}{Whether to drop genes that don't have human orthologues.}

\item{input_species}{The species that the `exp` dataset comes from
(to be used during the non-orthologues filtering step).}

\item{return_sce}{Whether to return the filtered results
as an expression matrix or a \pkg{SingleCellExperiment}.}

\item{verbose}{Whether to print messages (`T` or `F`).}

\item{...}{Additional arguments to be passed to the selected DGE method.}

\item{`level2annot}{Array of cell types, with each sequentially corresponding a column in the expression matrix}
}
\value{
exp Expression matrix with gene names as rownames,
or a \pkg{SingleCellExperiment} if \code{return_sce=T}.
}
\description{
\code{drop.uninformative.genes} drops genes from an SCT expression matrix if they do not significantly vary between any celltypes.
Makes this decision based on use of an ANOVA (implemented with Limma). If the F-statistic for variation amongst type2 annotations
is less than a strict p-threshold, then the gene is dropped.
}
\examples{
data("cortex_mrna")
cortex_mrna$exp = cortex_mrna$exp[1:300,] # Use only a subset of genes to keep the example quick
exp2 = drop.uninformative.genes(exp=cortex_mrna$exp, level2annot=cortex_mrna$annot$level2class)
}
