% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_controlled_bootstrap_geneset.r
\name{generate_controlled_bootstrap_geneset}
\alias{generate_controlled_bootstrap_geneset}
\title{generate_controlled_bootstrap_geneset}
\usage{
generate_controlled_bootstrap_geneset(
  hitGenes,
  sct_data,
  annotLevel,
  reps,
  controlledCT = NULL
)
}
\arguments{
\item{hitGenes}{Array of gene names. The target gene set.}

\item{sct_data}{The cell type data list (with specificity and mean_exp)}

\item{annotLevel}{The level of annotation in sct_data to analyse}

\item{reps}{The number of gene lists to sample}

\item{controlledCT}{Name of a celltype (from colnames of 
sct_data[[x]]$specificity).}
}
\value{
Matrix of genes (nrows=length(hitGenes),ncols=reps), where each 
column is a gene list
}
\description{
\code{generate_controlled_bootstrap_geneset} Used to generated celltype 
controlled bootstraped
}
\examples{
# See vignette
}
