% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exp_data_for_bootstrapped_genes.R
\name{get_exp_data_for_bootstrapped_genes}
\alias{get_exp_data_for_bootstrapped_genes}
\title{get_exp_data_for_bootstrapped_genes}
\usage{
get_exp_data_for_bootstrapped_genes(
  results,
  signif_res,
  sct_data,
  hits,
  combinedGenes,
  annotLevel,
  nReps = 100,
  as_sparse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{signif_res}{signif_res (#fix).}

\item{sct_data}{List generated using \link[EWCE]{generate_celltype_data}.}

\item{hits}{Gene hits.}

\item{combinedGenes}{Combined list of genes from \code{sct_data}, 
\code{hits}, and background \code{bg}.}

\item{annotLevel}{An integer indicating which level of \code{sct_data} to
analyse (\emph{Default: 1}).}

\item{verbose}{Print messages.}

\item{full_results}{full_results (#fix).}
}
\value{
exp_mats
}
\description{
Support function for 
\link[EWCE]{generate_bootstrap_plots_for_transcriptome}.
}
\keyword{internal}
