% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summed_proportions.r
\name{get_summed_proportions}
\alias{get_summed_proportions}
\title{Get summed proportions}
\usage{
get_summed_proportions(
  hitGenes,
  sct_data,
  annotLevel,
  reps,
  geneSizeControl,
  controlledCT = NULL,
  control_network = NULL
)
}
\arguments{
\item{hitGenes}{Array of gene names. The target gene set.}

\item{sct_data}{The cell type data list (with specificity and mean_exp)}

\item{annotLevel}{The level of annotation in sct_data to analyse}

\item{reps}{The number of gene lists to sample}

\item{geneSizeControl}{Boolean. Should the sampled gene lists control for 
GC content and transcript length?}

\item{controlledCT}{Name of a celltype (from colnames of 
sct_data[[x]]$specificity).}

\item{control_network}{If geneSizeControl=TRUE, then must provide the 
control network}
}
\value{
A list containing three data frames:
\itemize{
  \item \code{hit.cells}: vector containing the summed proportion of 
  expression in each cell type for the target list
  \item \code{bootstrap_data}: matrix in which each row represents the 
  summed proportion of expression in each cell type for one of the 
  random lists
  \item \code{controlledCT}: the controlled celltype (if applicable)
}
}
\description{
\code{get_summed_proportions} Given the target geneset, randomly sample 
genelists of equal length, obtain the specificity of these and then
obtain the mean specificity in each sampled list (and the target list)
}
\examples{
# See bootstrap.enrichment.test.r
}
