% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_with_bootstrap_distributions.R
\name{plot_with_bootstrap_distributions}
\alias{plot_with_bootstrap_distributions}
\title{Plot with bootstrap distributions}
\usage{
plot_with_bootstrap_distributions(
  exp_mats,
  cc,
  hit_exp,
  tag,
  listFileName,
  graph_theme,
  save_dir = file.path(tempdir(), paste0("BootstrapPlots", "_for_transcriptome")),
  height = 3.5,
  width = 3.5
)
}
\value{
Null result.
}
\description{
Plot results of \link[EWCE]{generate_bootstrap_plots_for_transcriptome}.
}
\keyword{internal}
