% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_genesize_control_network.r
\name{prepare_genesize_control_network}
\alias{prepare_genesize_control_network}
\title{Prepare genesize control network}
\usage{
prepare_genesize_control_network(
  hits,
  bg = NULL,
  numBOOT = 10000,
  sctSpecies = NULL,
  genelistSpecies = NULL,
  verbose = TRUE
)
}
\arguments{
\item{hits}{List of gene symbols containing the target gene list.
Will automatically be converted to human gene symbols
if \code{geneSizeControl=TRUE}.}

\item{bg}{List of gene symbols containing the background gene list
(including hit genes). If \code{bg=NULL},
 an appropriate gene background will be created automatically.
if \code{geneSizeControl=TRUE}.}

\item{numBOOT}{Number of gene lists to sample.}

\item{sctSpecies}{Species that \code{sct_data} came from
(no longer limited to just "mouse" and "human").}

\item{genelistSpecies}{Species that \code{hits} genes came from
(no longer limited to just "mouse" and "human").}

\item{verbose}{Print messages.}
}
\value{
A list containing three data frames:
\itemize{
  \item \code{hitGenes}: Array of HGNC symbols containing the hit genes.
  May be slightly reduced if gene length / GC content could not be found
  for all genes.
  \item \code{list_network}: The control gene lists as a data frame of HGNC
  symbols
}
}
\description{
\code{prepare_genesize_control_network} takes a gene list and finds
semi-randomly selected gene lists which are matched for gene length and
GC content.
}
\keyword{internal}
