% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_limma.r
\name{run_limma}
\alias{run_limma}
\title{Run DGE: \pkg{limma}}
\usage{
run_limma(exp, level2annot, mtc_method = "BH", verbose = TRUE, ...)
}
\arguments{
\item{exp}{Expression matrix with gene names as rownames.}

\item{level2annot}{Array of cell types, with each sequentially corresponding
a column in the expression matrix.}

\item{mtc_method}{Multiple-testing correction method used by DGE step.
See \link[stats]{p.adjust} for more details.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments to be passed to the selected DGE method.}
}
\value{
\code{limma} results.
}
\description{
Run Differential Gene Expression with \pkg{limma}.
}
\keyword{internal}
