% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scT_normalize.R
\name{scT_normalize}
\alias{scT_normalize}
\title{Normalize expression matrix}
\usage{
scT_normalize(exprs)
}
\description{
Normalize expression matrix by accounting for library size.
Uses \pkg{sctransform}.
}
