% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_delayed_array.R
\name{to_delayed_array}
\alias{to_delayed_array}
\title{Convert object to DelayedArray}
\usage{
to_delayed_array(exp, as_DelayedArray = TRUE, verbose = TRUE)
}
\arguments{
\item{exp}{Object.}

\item{as_DelayedArray}{Whether to convert \code{exp} to DelayedArray.}

\item{verbose}{Print messages.}
}
\description{
Convert a variety of object types to DelayedArray format.
}
\keyword{internal}
